+++ fdtd_1d_maxwell.h
@@ -1,5 +1,5 @@
 #ifndef FDTD_1D_MAXWELL_H
 #define FDTD_1D_MAXWELL_H
-void update_B(int Nx, double **Bx, double **By, double **Ez, double xi);
-void update_Dz(int Nx, double **Dz, double **Hx,  double **Hy, double source, double xi);
+void update_B(int Nx, float **Bx, float **By, float **Ez, float xi);
+void update_Dz(int Nx, float **Dz, float **Hx,  float **Hy, float source, float xi);
 #endif // FDTD_QD_MAXWELL_H
